add script to investigate model performance directly with ollama
The outcome of this is that even with RoPE working correctly,
small models get overwhelmed by the combination of RoPE and complex prompts.
So if we have longer output, we have to make our tasks simpler,
or split them up in small enough chunks so that RoPE does not play a role.