fix inner loop bug, add more coordinator messages
I switched the != to == which made it so build errors actually exit the inner loop when they should.
I added more coordinator messages especially to understand if the build error turns to progress
or not.
Switched the error.type check to after the success check cause when build succeeds error is none.