implement special handling of source attribute
This prevents us from running into
raise RuntimeError(f"Failed to retrieve build logs for {derivation_path}: {log_result.stderr}")
when the issue is actually with the source attribute.
This solution does not really allow for much iteration on the source attribute,
because it always treats errors on it as a regression and fixes to it should probably
have additional information in the form of the first packaging prompt.
We might even want to iterate on the source attribute ahead of the main
build iteration loop in a more tailored manner.
We can implement something better like that in the future.