fix: improve literal package search
It happens often the model specifies the "wrong" package set, for example, one that doesn't exist, or python3Packages, which it can use in packaging code but not during search.
With these changes, even when this happens, there's a lot more chances that what the model is searching for will be present in the results.
Searching for python3Packages set doesn't result in any packages, but then the remaining package slots in the results can be filled with approximate packages, for python312Packages, for example.
And other related "failing upwards" changes.