feat: add search_nixpkgs_manual_documentation tool to SEARCH_FUNCTIONS
This tool primarily serves to more directly give the model access to the nixpkgs reference manual documentation, currently specifically only on the language and frameworks section.
While previously there have been instances in which the model used file tools to search and read these documentation files, this tool reduces the number of tool calls used to discover them, and is a constant reminder that they exist and should be used when relevant.
When no exact match is found for a documentation file name (e.g. "buildPythonApplication" != "python"), the query is considered a keyword and it is searched across every documentation file. Whichever documentation file has the most matches for the keyword, is returned to the model.
This helps in situations where the model searches for a framework, or whatever else, that does not have its own dedicated documentation (e.g. "flutter").